+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/basicLogicalOperators.scala
@@ -660,10 +660,10 @@ object CTERelationDef {
  * @param statsOpt  The optional statistics inferred from the corresponding CTE definition.
  */
 case class CTERelationRef(
-  cteId: Long,
-  _resolved: Boolean,
-  override val output: Seq[Attribute],
-  statsOpt: Option[Statistics] = None) extends LeafNode with MultiInstanceRelation {
+    cteId: Long,
+    _resolved: Boolean,
+    override val output: Seq[Attribute],
+    statsOpt: Option[Statistics] = None) extends LeafNode with MultiInstanceRelation {
 
   final override val nodePatterns: Seq[TreePattern] = Seq(CTE)
